Fix list behavior when loading file

+++ lib/data/data_source/document_remote_data_source.dart
@@ -4,6 +4,6 @@ abstract class IDocumentRemoteDataSource {
   Future<void> loadDocument({
     required String url,
     required String filePath,
-    required StreamController<int> progressStream,
+    required StreamController<double> progressStream,
   });
 }
