+++ lib/domain/model/document.dart
@@ -12,7 +12,7 @@ class Document with _$Document {
     required String name,
     required String url,
     required DocumentStatus status,
-    StreamController<int>? downloadProgressStream,
+    StreamController<double>? downloadProgressStream,
     String? filePath,
   }) = _Document;
 }
